+++ packages/pkg-core/src/interfaces/lifecycle/synchronise-workspace-lifecycle-base.ts
@@ -110,6 +110,7 @@ export class SynchroniseWorkspaceLifecycleBase<
     await SynchroniseWorkspaceLifecycleBase.hooks.beforeSynchroniseWorkspace.callBailAsync(
       { options }
     );
+
     const res =
       await SynchroniseWorkspaceLifecycleBase.hooks.synchroniseWorkspace.callBailAsync(
         { options, strict: true }