+++ packages/tests/pkg-tests-specs-fixtures/src/stage2/configure/configure-workspace-stage2.spec.ts
@@ -15,6 +15,7 @@ describe("configure workspace", () => {
         testContext.testEnvVars.workspacePath
       );
 
+      // @ts-ignore
       let expectConfig = resCopy.res.expectUtil.createConfig();
 
       const res = await testContext.command(