+++ packages/tests/pkg-tests-verdaccio/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @business-as-code/tests-verdaccio
 
+## 0.0.0-bollards-20230810141915
+
+### Patch Changes
+
+- Updated dependencies
+  - @business-as-code/core@0.0.0-bollards-20230810141915
+  - @business-as-code/fslib@0.0.0-bollards-20230810141915
+
 ## 0.0.0-latest-20230716081621
 
 ### Patch Changes